refactor/recipe viewer + minor ui improvements
